backend/requirements/production.txt
# Production requirements for ZentraQMS Backend
# These packages are needed in production environment

-r base.txt

# Production-specific packages
gunicorn==21.2.0
django-redis==5.4.0

# Security
django-csp==3.7.0
django-security==0.17.0

# Monitoring and logging
sentry-sdk==1.38.0
django-health-check==3.17.0

# Performance
django-cachalot==2.6.1

# Task queue (production-ready)
celery[redis]==5.3.4
flower==2.0.1

# File storage (if using cloud storage)
django-storages==1.14.2
boto3==1.34.0

# Environment
python-decouple==3.8